refactor: move some workflow types to the dedicated types directory
