+++ packages/twenty-front/src/modules/workflow/types/WorkflowDiagram.ts
@@ -1,0 +1,10 @@
+export type WorkflowDiagramStepNodeData = {
+  nodeType: 'trigger' | 'condition' | 'action';
+  label: string;
+};
+
+export type WorkflowDiagramCreateStepNodeData = Record<string, never>;
+
+export type WorkflowDiagramNodeData =
+  | WorkflowDiagramStepNodeData
+  | WorkflowDiagramCreateStepNodeData;
