+++ packages/twenty-front/src/pages/workflows/nodes/base.tsx
@@ -2,17 +2,6 @@ import styled from '@emotion/styled';
 
 import { Handle } from '@xyflow/react';
 
-export type WorkflowDiagramStepNodeData = {
-  nodeType: 'trigger' | 'condition' | 'action';
-  label: string;
-};
-
-export type WorkflowDiagramCreateStepNodeData = Record<string, never>;
-
-export type WorkflowDiagramNodeData =
-  | WorkflowDiagramStepNodeData
-  | WorkflowDiagramCreateStepNodeData;
-
 export const StyledTargetHandle = styled(Handle)`
   visibility: hidden;
 `;
